requirements.txt
aenum==3.1.8
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
dill==0.3.4
h11==0.13.0
idna==3.3
iniconfig==1.1.1
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.1
pytest-reportportal==5.0.12
python-dotenv==0.19.2
reportportal-client==5.1.0
requests==2.27.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
termcolor==1.1.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.5.3
wsproto==1.0.0
